Pokecards v0.1.0 Pokecards

Provides functions to create and handle a deck of pokemons

Functions

constains?(deck, pokemon)
(since 0.1.0)

Verifies if a deck constains a pokemon

Examples

iex> deck = Pokecards.create_deck
iex> Pokecards.constains?(deck, "Name: Piplup, Ability: Torrent, Damage: 50")
true

create_deck()
(since 0.1.0)

Creates a deck of pokemons

Examples

iex> deck = Pokecards.create_deck
iex> deck

create_hand(hand_size)
(since 0.1.0)

Returns a hand of a pokemon deck based on a hand_size

Examples

iex> hand = Pokecards.create_hand(5)
iex> length(hand) == 5
true

deal(deck, hand_size)
(since 0.1.0)

Deal a deck of pokemons by a given a hand_size

Examples

iex> deck = Pokecards.create_deck
iex> { hand, _rest_of_hand } = Pokecards.deal(deck, 2)
iex> length(hand) == 2
true

load(deckname)
(since 0.1.0)

Loads a deck with the given deckname from the filesystem

Examples

iex> deck = Pokecards.load("test")
iex> deck

save(deckname, deck)
(since 0.1.0)

Saves a deck with a given deckname to the filesystem

Returns :ok

Examples

iex> deck = Pokecards.create_deck
iex> Pokecards.save("cardname", deck)
:ok

shuffle(deck)
(since 0.1.0)

Shuffles a deck of pokemons

Examples

iex> deck = Pokecards.create_deck
iex> deck = Pokecards.shuffle(deck)
iex> deck